place BlogPostContain in separate directory
add styles.js file
