+++ pages/blog/index.js
@@ -1,5 +1,5 @@
 import { useState } from 'react';
-import BlogPostsContainer from '@/components/blog/BlogPostsContainer';
+import BlogPostsContainer from '@/components/blog/BlogPostsContainer/BlogPostsContainer';
 import Container from '@/components/containers/Container';
 import SearchBar from '@/components/blog/SearchBar';
 import Title from '@/components/snippets/Title';
